+++ tweetshake/LoginViewController.swift
@@ -27,8 +27,8 @@ class LoginViewController: UIViewController {
         }
     }
 
-    override func viewDidDisappear(animated: Bool) {
-        super.viewDidDisappear(animated)
+    override func viewWillDisappear(animated: Bool) {
+        super.viewWillDisappear(animated)
         isLoginViewVisible = false
         logInButton?.removeFromSuperview()
         logInButton = nil